fix(bundle): move check-bundle-size tests alongside action scripts
Split the test file: check-bundle-size tests move to the action
directory, update-bundle-sizes tests stay in scripts/. Add .github/
actions glob to vitest scripts config so the moved tests are picked up.
